+++ benches/hybrid_cache_soak.rs
@@ -112,7 +112,6 @@ impl SoakConfig {
         StaticConfig::new(self.capacity_bytes)
             .with_region_size(32 * MIB as u64)
             .with_expected_entries(self.key_count)
-            .with_write_shards(self.shards)
     }
 
     fn runtime_config(&self) -> RuntimeConfig {
@@ -121,6 +120,7 @@ impl SoakConfig {
             .with_io_mode(self.io_mode)
             .with_read_io_workers(self.read_io_workers)
             .with_write_io_workers(self.write_io_workers)
+            .with_write_shards(self.shards)
             .with_l1_capacity(self.memory_bytes)
             .with_memory_limit(self.memory_limit_bytes)
             .with_statistics(true)